elisp: Fix all snippets with a literal &optional in them
The literal &optional in the snippets makes for syntax errors if they
are used. Since this makes for more cleanup than just manually writing whatever
the snippets would expand to, this made all of these pretty useless.

Also, many of the snippets were barely templates - they inserted the
boilerplate as suggested by the docstring of their functions (hence the
&optional, one could guess). This fixes that by making them into
tabbable snippet areas. Everything that is optional is preceded by a
single space, so skipping over them with `C-d` leaves the snippet in a
desirable state, without trailing spaces.