Remove paren-wrapping snippets from emacs-lisp-mode
There are several snippets whose expansion is nothing more than

    (KEY $0)

where KEY is the snippet's #key field.  These snippets do nothing but
save you typing in the parentheses.  Using a package like paredit, or
smartparens would be more suited to this.  There are currently a lot
of snippets in emacs-lisp-mode, which makes browsing through the list
more difficult, so these least useful ones should be removed.