org-mode/src: Suppress indentation
Before this commit, trying to use the org-mode/src snippet with
org-src-tab-acts-natively set to t would insert the snippet, print an error
message in the echo area: "No such language mode: nil-mode", and leave
point at the beginning of the first line of the snippet.

After this commit, the snippet is expanded with no error message and with
point in the correct position.

* snippets/org-mode/src: Set org-src-tab-acts-natively to nil to prevent
Org-Mode from trying to indent the block with nil major mode.