remove/rename a lot of duplicated snippets, and add tests
There were currently many snippets with the same name in the same
mode, which means they could never be activated.
With this PR we make a pretty big cleanup of all the ones that were
unusable, by removing or renaming them.
Also add a couple of tests to ensure more consistency across the board.