Merge elpy snippets (again) (#465)
* elpy-initialize-variables: Add the Elpy snippet directory.

Elpy now ships with a few YASnippet snippets for Python.

Fixes #43.

* Small rework of the snippets.

The enter and exit snippets now conform to the yasnippet-snippets
repository (mostly), while super now makes use of the .yas-setup.el
mechanism and is seriously improved. There's a new snippet for
__init__ methods, too, which automatically assigns arguments to
instance variables.

* Rework and expand the snippet collection.

* Add __bool__ snippet.

* updated yasnippet variable 'text' to 'yas-text'

* Return correct form of "super" depending on Python version

Previously when using the yasnippet "super", it will always return
"(class, arg).method" which was the Py2 way of doing it. In Py3, you
could simply do "().method" and Python will understand.

* Move yasnippet-snippets onto their own branch

* PR #259 Change keys for __foo__ from "foo" to "_foo"

Thanks to galaunay for the idea to use this to "avoid unnecessary
keyspace pollution."

* Fix bug due to passage from elpy to yasnippet

* Use snippet expansion for foo -> __foo__

* Fix bug when specifying arguments on several lines

* Allow detection of Python 3 on PEP 394-compliant systems.

Please note that python-interpreter must be set to "python3" for this
to function correctly.

As a side-effect it also allows the use of Python 2 on systems that
install version 3 to /usr/bin/python and version 2 to
/usr/bin/python2.

* Change elpy prefixes to yas-snips

Thanks to Andrea Crotti for the review and suggestion.

* Add newline at end of file for all files that were missing one.

* Make python snippets able to handle arguments with type annotations.

* Use start-point from yas--apply-transform in snippet-init-assignment

yas-snips-snippet-init-assignments should use values provided by
yas--apply-transform instead of grabbing them itself.

Thanks to npostavs for finding this issue.  See review discussion at
PR #278 for more info (search for "yas--apply-transform").

* Removing Python 2 snippets, it's EOL since 2020-01-01.

* Some consistenty in group names.

* Removing a duplicate.

* Owner is documented as optional.

cf. https://docs.python.org/3/reference/datamodel.html#object.__get__

* Resolve a conflict between _getattr and _getattribute

* Deduplicate: it already exists in master (named pdb).

* Dedup: already exists as __repr__

* Dedup: already exists as __str__

* fix: try is already used by ./try

* fix misleading key.

* Consistency with other dunders.

* Consistency with other naming.

* Removing super, we don't have all needed elisp yet.

* Deduplicate two keys.

* s not needed.

* The hook does not works for me.

* Add few missing newlines at end of file.

---------

Co-authored-by: Jorgen Schaefer <forcer@forcix.cx>
Co-authored-by: Jorgen Schaefer <contact@jorgenschaefer.de>
Co-authored-by: Daniel Wu <dawu@wgen.net>
Co-authored-by: Daniel Gopar <gopardaniel@yahoo.com>
Co-authored-by: Nicholas D Steeves <nsteeves@gmail.com>
Co-authored-by: galaunay <gaby.launay@tutanota.com>