Create `nix-ts-mode` sub-directory; Add additional snippets to `nix-mode` (#515)
* nix-ts-mode: New snippet directory.

* nix-mode: pkg_mkshell: New snippet

This snippet is designed to quickly write out a `shell.nix` for Nix.

* nix-mode: doCheck: New snippet.

Adds a snippet to add the doCheck attribute to a package derivation
that uses the current standard: a condition for cross-compilation or
false.

* nix-mode: passthru_update: New snippet.

This adds a snippet for adding the `passthru.updateScript` attribute
used in many nixpkgs derivations.